Moved gossip codec into its own package
